Moved AdditionalColumns to its own package
